wip: implementing new types

+++ clickhouse-arrow/src/native/values.rs
@@ -287,16 +287,16 @@ impl Value {
             Value::UInt256(_) => Type::UInt256,
             Value::Float32(_) => Type::Float32,
             Value::Float64(_) => Type::Float64,
-            Value::Decimal32(p, _) => Type::Decimal32(*p),
-            Value::Decimal64(p, _) => Type::Decimal64(*p),
-            Value::Decimal128(p, _) => Type::Decimal128(*p),
-            Value::Decimal256(p, _) => Type::Decimal256(*p),
+            Value::Decimal32(p, _) => Type::Decimal32(*p as u8),
+            Value::Decimal64(p, _) => Type::Decimal64(*p as u8),
+            Value::Decimal128(p, _) => Type::Decimal128(*p as u8),
+            Value::Decimal256(p, _) => Type::Decimal256(*p as u8),
             Value::String(_) => Type::String,
             Value::Uuid(_) => Type::Uuid,
             Value::Date(_) => Type::Date,
             Value::Date32(_) => Type::Date32,
             Value::DateTime(time) => Type::DateTime(time.0),
-            Value::DateTime64(x) => Type::DateTime64(x.2, x.0),
+            Value::DateTime64(x) => Type::DateTime64(x.2 as u8, x.0),
             Value::Enum8(_, i) => Type::Enum8(vec![(String::new(), *i)]),
             Value::Enum16(_, i) => Type::Enum16(vec![(String::new(), *i)]),
             Value::Array(x) => {
@@ -434,9 +434,11 @@ impl fmt::Display for Value {
                 write!(f, "')")
             }
             Value::DateTime64(datetime) => {
-                let chrono_date: chrono::DateTime<Tz> =
-                    FromSql::from_sql(&Type::DateTime64(datetime.2, datetime.0), self.clone())
-                        .map_err(|_| fmt::Error)?;
+                let chrono_date: chrono::DateTime<Tz> = FromSql::from_sql(
+                    &Type::DateTime64(datetime.2 as u8, datetime.0),
+                    self.clone(),
+                )
+                .map_err(|_| fmt::Error)?;
                 let string = chrono_date.to_rfc3339_opts(SecondsFormat::AutoSi, true);
                 write!(f, "parseDateTime64BestEffort('")?;
                 escape_string(f, &string)?;
